+++ lanzou/gui/workers/login.py
@@ -1,6 +1,6 @@
 from PyQt5.QtCore import QThread, pyqtSignal, QMutex
-from lanzou.api.utils import logger
 from lanzou.api import LanZouCloud
+from lanzou.debug import logger
 
 
 class LoginLuncher(QThread):
