+++ lanzou/gui/workers/rename.py
@@ -1,7 +1,7 @@
 from time import sleep
 from PyQt5.QtCore import QThread, pyqtSignal, QMutex
-from lanzou.api.utils import logger
 from lanzou.api import LanZouCloud
+from lanzou.debug import logger
 
 
 class RenameMkdirWorker(QThread):
