+++ lanzou/gui/workers/rm.py
@@ -1,5 +1,5 @@
 from PyQt5.QtCore import QThread, pyqtSignal, QMutex
-from lanzou.api.utils import logger
+from lanzou.debug import logger
 
 
 class RemoveFilesWorker(QThread):
