+++ main.py
@@ -9,7 +9,7 @@ from lanzou.gui.gui import MainWindow
 
 if __name__ == "__main__":
     app = QApplication(sys.argv)
-    app.setWindowIcon(QIcon("./lanzou/gui/src/lanzou_logo2.png"))
+    app.setWindowIcon(QIcon("./src/lanzou_logo2.png"))
     form = MainWindow()
     form.show()
     sys.exit(app.exec())
